Define domain for meetup attendance
The main UI, for the time being, is going to be the database.

This app serves as the driver of the data schema and later as a set
API for a future frontend, once users feel confident enough about
the most common queries they need to visualize and data they need
to insert.